二号小论文/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "your_dl_project"
version = "0.1.0"
authors = [
  { name="Your Name", email="you@example.com" },
]
description = "A deep learning project with structured architecture."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "scikit-learn>=1.0.0",
    "pyyaml>=5.4.0",
    # Uncomment the framework you plan to use:
    # "torch>=1.10.0",
    # "tensorflow>=2.8.0",
]

[project.urls]
"Homepage" = "https://github.com/your-username/your_dl_project"
"Bug Tracker" = "https://github.com/your-username/your_dl_project/issues"
